+++ mizani/__init__.py
@@ -1,4 +1,4 @@
-from ._version import get_versions
+from . import _version
 
-__version__ = get_versions()['version']
-del get_versions
+__version__ = _version.get_versions()['version']
+del _version