+++ doc/changelog.rst
@@ -18,6 +18,7 @@ API Changes
 v0.7.3
 ------
 *(2020-10-29)*
+
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4633364.svg
    :target: https://doi.org/10.5281/zenodo.4633364
 
@@ -30,6 +31,7 @@ Bug Fixes
 v0.7.2
 ------
 *(2020-10-29)*
+
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4633357.svg
    :target: https://doi.org/10.5281/zenodo.4633357